backend/requirements.txt
# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LLM
openai>=1.30.0

# Vector DB + Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# PDF parsing
PyMuPDF>=1.23.0

# Skin model
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Image
opencv-python-headless>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data
kaggle>=1.5.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Config
python-dotenv>=1.0.0
